build publish latest version to npm
