+++ lib/index.js
@@ -33,10 +33,6 @@ var _throttle = require('./utils/throttle');
 
 var _throttle2 = _interopRequireDefault(_throttle);
 
-var _decorator = require('./decorator');
-
-var _decorator2 = _interopRequireDefault(_decorator);
-
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
 function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
@@ -81,23 +77,18 @@ var checkOverflowVisible = function checkOverflowVisible(component, parent) {
   var parentHeight = void 0;
   var parentWidth = void 0;
 
-  var parentWidth = void 0;
-  var parentLeft = void 0;
-
   try {
     var _parent$getBoundingCl = parent.getBoundingClientRect();
 
     parentTop = _parent$getBoundingCl.top;
     parentLeft = _parent$getBoundingCl.left;
     parentHeight = _parent$getBoundingCl.height;
-
     parentWidth = _parent$getBoundingCl.width;
-    parentLeft = _parent$getBoundingCl.left;
-
   } catch (e) {
     parentTop = defaultBoundingClientRect.top;
     parentLeft = defaultBoundingClientRect.left;
     parentHeight = defaultBoundingClientRect.height;
+    parentWidth = defaultBoundingClientRect.width;
   }
 
   var windowInnerHeight = window.innerHeight || document.documentElement.clientHeight;
@@ -109,16 +100,10 @@ var checkOverflowVisible = function checkOverflowVisible(component, parent) {
   var intersectionHeight = Math.min(windowInnerHeight, parentTop + parentHeight) - intersectionTop; // height
   var intersectionWidth = Math.min(windowInnerWidth, parentLeft + parentWidth) - intersectionLeft; // width
 
-  var intersectionLeft = Math.max(parentLeft, 0)
-  var intersectionWidth = Math.min(windowInnerWidth, parentLeft + parentWidth) - intersectionLeft;
-  
   // check whether the element is visible in the intersection
   var top = void 0;
   var left = void 0;
   var height = void 0;
-  var width = void 0;
-
-  var left = void 0;
   var width = void 0;
 
   try {
@@ -127,13 +112,12 @@ var checkOverflowVisible = function checkOverflowVisible(component, parent) {
     top = _node$getBoundingClie.top;
     left = _node$getBoundingClie.left;
     height = _node$getBoundingClie.height;
-
     width = _node$getBoundingClie.width;
-    left = _node$getBoundingClie.left;
   } catch (e) {
     top = defaultBoundingClientRect.top;
     left = defaultBoundingClientRect.left;
     height = defaultBoundingClientRect.height;
+    width = defaultBoundingClientRect.width;
   }
 
   var offsetTop = top - intersectionTop; // element's top relative to intersection
@@ -158,32 +142,21 @@ var checkNormalVisible = function checkNormalVisible(component) {
   var top = void 0;
   var elementHeight = void 0;
 
-  var left = void 0;
-  var elementWidth = void 0;
-
   try {
     var _node$getBoundingClie2 = node.getBoundingClientRect();
 
     top = _node$getBoundingClie2.top;
     elementHeight = _node$getBoundingClie2.height;
-
-    left = _node$getBoundingClie2.left;
-    elementWidth = _node$getBoundingClie2.width;
-
   } catch (e) {
     top = defaultBoundingClientRect.top;
     elementHeight = defaultBoundingClientRect.height;
-
-    left = defaultBoundingClientRect.left;
-    elementWidth = defaultBoundingClientRect.width;
   }
 
   var windowInnerHeight = window.innerHeight || document.documentElement.clientHeight;
-  var windowInnerWidth = window.innerWidth || document.documentElement.clientWidth;
 
   var offsets = Array.isArray(component.props.offset) ? component.props.offset : [component.props.offset, component.props.offset]; // Be compatible with previous API
-  
-  return (top - offsets[0] <= windowInnerHeight && top + elementHeight + offsets[1] >= 0) && (left - offsets[0] <= windowInnerWidth && left + elementWidth + offsets[1] >= 0);
+
+  return top - offsets[0] <= windowInnerHeight && top + elementHeight + offsets[1] >= 0;
 };
 
 /**
@@ -197,7 +170,7 @@ var checkVisible = function checkVisible(component) {
   if (!(node instanceof HTMLElement)) {
     return;
   }
-  
+
   var parent = (0, _scrollParent2.default)(node);
   var isOverflow = component.props.overflow && parent !== node.ownerDocument && parent !== document && parent !== document.documentElement;
   var visible = isOverflow ? checkOverflowVisible(component, parent) : checkNormalVisible(component);
@@ -345,7 +318,7 @@ var LazyLoad = function (_Component) {
         listeners.splice(index, 1);
       }
 
-      if (listeners.length === 0) {
+      if (listeners.length === 0 && typeof window !== 'undefined') {
         (0, _event.off)(window, 'resize', finalLazyLoadHandler, passiveEvent);
         (0, _event.off)(window, 'scroll', finalLazyLoadHandler, passiveEvent);
       }
@@ -384,6 +357,41 @@ LazyLoad.defaultProps = {
   unmountIfInvisible: false
 };
 
-var lazyload = exports.lazyload = _decorator2.default;
+var getDisplayName = function getDisplayName(WrappedComponent) {
+  return WrappedComponent.displayName || WrappedComponent.name || 'Component';
+};
+
+var decorator = function decorator() {
+  var options = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : {};
+  return function lazyload(WrappedComponent) {
+    return function (_Component2) {
+      _inherits(LazyLoadDecorated, _Component2);
+
+      function LazyLoadDecorated() {
+        _classCallCheck(this, LazyLoadDecorated);
+
+        var _this2 = _possibleConstructorReturn(this, (LazyLoadDecorated.__proto__ || Object.getPrototypeOf(LazyLoadDecorated)).call(this));
+
+        _this2.displayName = 'LazyLoad' + getDisplayName(WrappedComponent);
+        return _this2;
+      }
+
+      _createClass(LazyLoadDecorated, [{
+        key: 'render',
+        value: function render() {
+          return _react2.default.createElement(
+            LazyLoad,
+            options,
+            _react2.default.createElement(WrappedComponent, this.props)
+          );
+        }
+      }]);
+
+      return LazyLoadDecorated;
+    }(_react.Component);
+  };
+};
+
+exports.lazyload = decorator;
 exports.default = LazyLoad;
 exports.forceCheck = lazyLoadHandler;
