+++ lib/utils/scrollParent.js
@@ -33,7 +33,7 @@ exports.default = function (node) {
       continue;
     }
 
-    if (overflowRegex.test(overflow) || overflowRegex.test(overflowX) || overflowRegex.test(overflowY)) {
+    if (overflowRegex.test(overflow) && overflowRegex.test(overflowX) && overflowRegex.test(overflowY)) {
       return parent;
     }
