+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-lazyload",
-  "version": "2.5.0",
+  "version": "2.6.2",
   "description": "Lazyload your components, images or anything where performance matters.",
   "main": "lib/index.js",
   "scripts": {
